+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Add "--no-header" command line argument.
+- Add "NoSpaceTable" output style.
 
 ## [0.4.5] - 2023-05-29